Cleaned up main.py and organized the main UI code.
